% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpe_ens_2018.R
\docType{data}
\name{bpe_ens_2018}
\alias{bpe_ens_2018}
\title{Base de données Base Permanentes des Équipements}
\format{
Une table de 1 035 564 lignes et 7 variables
\describe{
\item{REG}{Région d’implantation de l’équipement}
\item{DEP}{DEP Département d’implantation de l’équipement}
\item{DEPCOM}{Département et commune d’implantation de l’équipement}
\item{DCIRIS}{Département, commune et IRIS d’implantation de l’équipement}
\item{AN}{Année}
\item{TYPEQU}{Type d'équipement}
\item{NB_EQUIP}{Nombre d’équipements}
}
}
\source{
\url{https://www.insee.fr/fr/statistiques/fichier/3568629/bpe18_ensemble_csv.zip}
}
\usage{
bpe_ens_2018
}
\description{
La base permanente des équipements (BPE) est une source statistique qui fournit le niveau d'équipements et de services rendus à la population sur un territoire. Les résultats sont proposés sous forme de bases de données dans différents formats et pour deux niveaux géographiques : les communes et les Iris.
}
\keyword{datasets}
