% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filosofi_epci_2016.R
\docType{data}
\name{filosofi_epci_2016}
\alias{filosofi_epci_2016}
\title{Données sur les revenus et la pauvreté en 2016, niveau EPCI}
\format{

}
\source{
\url{https://www.insee.fr/fr/statistiques/3560121}
}
\usage{
filosofi_epci_2016
}
\description{
Dispositif Fichier localisé social et fiscal (Filosofi)
}
\keyword{datasets}
