% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utile.R
\name{infosDonnees}
\alias{infosDonnees}
\title{Recherche ligne d'informations dans liste_donnees}
\usage{
infosDonnees(donnees, date = NULL)
}
\arguments{
\item{donnees}{le nom court des données que l'on souhaite télécharger sur le site de l'Insee. La description complète des données associés à ce nom figure dans la table \code{\link{liste_donnees}}. Insensible à la casse.}

\item{date}{optionnel : le millésime des données si nécessaire. Peut prendre le format YYYY ou encore DD/MM/YYYY ; dans le dernier cas, on prendra le premier jour de la période de référence. Spécifier \code{"dernier"} sélectionne automatiquement le millésime le plus récent.}
}
\value{
Une unique ligne de \code{liste_donnees} (sous forme de list).
}
\description{
À partir de l'identifiant et d'une éventuelle date, isole la ligne de
\code{\link{liste_donnees}} correspondante.
}
\details{
La fonction retourne une erreur si aucune ligne ne correspond. Elle suggère
dans ce cas des noms d'identifiants proches et les millésimes disponibles si
l'année doit être spécifiée.
}
