% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utile.R
\name{ld_complet}
\alias{ld_complet}
\title{Charge le catalogue complet (statique + Melodi) de manière lazy}
\usage{
ld_complet()
}
\value{
un objet list (le catalogue complet)
}
\description{
À la première utilisation, récupère le catalogue Melodi depuis le SSP Cloud
et le fusionne avec le catalogue statique empaqueté dans le package.
Le résultat est mis en cache pour les appels suivants.
}
\keyword{internal}
