% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utile.R
\name{listerDonnees}
\alias{listerDonnees}
\title{Fournit la liste de données disponibles (données internes au package + requête dans le catalogue Melodi)}
\usage{
listerDonnees()
}
\value{
Une liste issue d'une fusion éventuelle (sous condition d'accès à Melodi) de la liste \code{ld} et de l'équivalent sous la catalogue Mélodi.
}
\description{
La fonction retourne une liste, \code{ld} seulement si pas d'accès Internet, et \code{ld} fusionné avec le catalogue Melodi sinon.
}
\keyword{internal}
