% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utile.R
\name{listToDf}
\alias{listToDf}
\title{Transforme une liste (typiquement \code{ld}) en data.frame}
\usage{
listToDf(liste, vars = NULL)
}
\arguments{
\item{liste}{un objet list à convertir en data.frame}

\item{vars}{optionnel, un vecteur de variables à récupérer}
}
\value{
Une data.frame
}
\description{
Objectif : ne pas créer une dépendance à \code{dplyr} et la fonction \code{bind_rows}, bien que du point de vue computationel, la fonction créée ici soit bien moins performante.
}
\keyword{internal}
