% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/millesimesDisponibles.R
\name{millesimesDisponibles}
\alias{millesimesDisponibles}
\title{Lister les différents millésimes pour une source donnée}
\usage{
millesimesDisponibles(donnees)
}
\arguments{
\item{donnees}{le nom court des données que l'on souhaite télécharger sur le site de l'Insee. La description complète des données associés à ce nom figure dans la table \code{\link{liste_donnees}}. Insensible à la casse.}
}
\value{
un vecteur contenant l'ensemble des millésimes disponibles pour une source de données.
}
\description{
Lister les différents millésimes pour une source donnée
}
\examples{
millesimesDisponibles("RP_LOGEMENT")
}
