% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utile.R
\name{recupererMelodi}
\alias{recupererMelodi}
\title{recupererMelodi ----------------------------------------------------}
\usage{
recupererMelodi(url)
}
\arguments{
\item{url}{le lien vers le json généré sur le SSP Cloud}
}
\value{
un objet list correspondant à la structure du json
}
\description{
Objectif : récupérer sous forme de liste le catalogue de Mélodi traité sur le SSP Cloud
}
\keyword{internal}
