% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utile.R
\name{retourneLangueMelodi}
\alias{retourneLangueMelodi}
\title{Fournit pour une liste de libelles, la langue pour chacun des éléments de la liste (adhérent à la structure du catalogue de Melodi)}
\usage{
retourneLangueMelodi(listeLibelles)
}
\value{
Un vecteur du type \code{c('fr', 'en')}
}
\description{
La fonction retourne un vecteur de codes langues
}
\keyword{internal}
