% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utile.R
\name{retourneLibelleMelodi}
\alias{retourneLibelleMelodi}
\title{Fournit le libellé correspondant à la langue désirée (adhérent à la structure du catalogue de Melodi)}
\usage{
retourneLibelleMelodi(listeLibelles, langue)
}
\value{
Une chaîne de caractères
}
\description{
La fonction retourne une chaîne de caractères dans la langue désirée, sélectionnée dans une liste de libellés
}
\keyword{internal}
