% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/telechargerDonnees.R
\name{telechargerDonnees}
\alias{telechargerDonnees}
\title{Téléchargement des données sur le site de l'Insee}
\usage{
telechargerDonnees(
  donnees,
  date = NULL,
  telDir = getOption("doremifasol.telDir"),
  argsApi = NULL,
  vars = NULL,
  force = FALSE,
  ...
)
}
\arguments{
\item{donnees}{le nom des données que l'on souhaite télécharger sur le site de l'Insee, que l'on peut retrouver dans la table \code{\link{liste_donnees}}.}

\item{date}{optionnel : le millésime des données si nécessaire. Peut prendre le format YYYY ou encore DD/MM/YYYY ; dans le dernier cas, on prendra le premier jour de la période de référence.}

\item{telDir}{optionnel : le dossier dans lequel sont téléchargées les données brutes. Par défaut, la valeur définie par \code{options(doremifasol.telDir = ...)}. Si l'utilisateur n'a pas défini cette valeur au préalable, un dossier temporaire de cache.}

\item{argsApi}{optionnel : dans le cas où c'est une API REST qui est utilisée, il est possible de spécifier des paramètres spécifiques à cette API de manière à collecter l'information désirée.}

\item{vars}{optionnel : un vecteur pour spécifier les variables à importer. Utile pour les données massives difficiles à charger en mémoire, voir section \emph{Details}.}

\item{force}{forcer le téléchargement, même si le fichier a déjà été téléchargé (et est identique).}

\item{...}{paramètres additionnels relatifs à l'importation des données}
}
\value{
un objet tibble contenant les données téléchargées sur le site de l'Insee.
}
\description{
Téléchargement des données sur le site de l'Insee
}
\details{
La fonction permet de télécharger les données disponibles sur le site de l'Insee sous format csv, xls ou encore xlsx. Les données mises à disposition sont en général des tables de taille raisonnable, qui peuvent être chargées sans problème en mémoire sur un large spectre de machines. Néanmoins, pour certaines données (telles celles du Recensement de Population ou encore SIRENE), les données sont très volumineuses et exigent donc des machines très performantes. L'utilisateur a donc la possibilité de choisir les variables qui l'intéressent et de ne charger que ces dernières en mémoire, de manière à être parcimonieux.
}
\examples{
\dontrun{
bpe_ens_2018 <- telechargerDonnees(donnees = "BPE_ENS")

rp_log <- telechargerDonnees("RP_LOGEMENT", date = "2016", vars = c("COMMUNE", "IPONDL", "CATL"))
}
}
