% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/telechargerFichier.R
\name{telechargerFichier}
\alias{telechargerFichier}
\title{Téléchargement des données sur le site de l'Insee}
\usage{
telechargerFichier(donnees, date = NULL, telDir = NULL, argsApi = NULL)
}
\arguments{
\item{donnees}{le nom des données que l'on souhaite télécharger sur le site de l'Insee, que l'on peut retrouver dans la table ([liste_donnees])}

\item{date}{optionnel : le millésime des données si nécessaire. Peut prendre le format YYYY ou encore DD/MM/YYYY ; dans le dernier cas, on prendra le premier jour de la période de référence.}

\item{telDir}{optionnel : le dossier dans lequel sont téléchargées les données brutes. Par défaut, un dossier temporaire de cache.}

\item{argsApi}{optionnel : dans le cas où c'est une API REST qui est utilisée, il est possible de spécifier des paramètres spécifiques à cette API de manière à collecter l'information désirée. Cf. [@details ].}
}
\value{
une liste contenant le résultat du téléchargement et les informations pour l'importation des données en R.
}
\description{
Téléchargement des données sur le site de l'Insee
}
\details{
La fonction permet de télécharger les données disponibles sur le site de l'Insee sous format csv, xls ou encore xlsx.
}
\examples{
\dontrun{
dl_bpe <- telechargerFichier(donnees = "BPE_ENS")

dl_rplog <- telechargerFichier("RP_LOGEMENT", date = "2016")
}
}
