% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/millesimesDisponibles.R
\name{millesimesDisponibles}
\alias{millesimesDisponibles}
\title{Lister les différents millésimes pour une source donnée}
\usage{
millesimesDisponibles(donnees)
}
\arguments{
\item{donnees}{une chaîne de caractères qui identifie la source de données}
}
\value{
un vecteur contenant l'ensemble des millésimes disponibles pour une source de données.
}
\description{
Lister les différents millésimes pour une source donnée
}
\examples{
millesimesDisponibles("RP_LOGEMENT")
}
