% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sirets_successeurs.R
\name{sirets_successeurs}
\alias{sirets_successeurs}
\title{Liens de succession d'une liste d'établissements}
\usage{
sirets_successeurs(sirets)
}
\arguments{
\item{sirets}{sirets pour lesquel rechercher les successeurs (un vecteur caractère)}
}
\value{
Un data.frame agrégeant les résultats pour chaque siret (vide si
aucun des établissements n'a de successeur).
}
\description{
Télécharge sur l'API Sirene l'ensemble des successeurs d'une liste
d'établissements.
}
\details{
Cette fonction appelle \code{\link{telechargerDonnees}} pour chacun des établissements
passés en paramètre (de manière optimisée).

Les données téléchargées au format "json" sont enregistrées dans un dossier
temporaire.
}
\section{Authentification API Sirene}{

Comme toutes les fonctions reposant sur l'\href{https://api.insee.fr}{API} Sirene,
cette fonction nécessite une clé d'application et un secret associé pour
pouvoir générer un jeton d'accès. Ces informations sont à passer sous forme
de variables d'envionnement.

Renseigner pour cela \code{INSEE_APP_KEY} et \code{INSEE_APP_SECRET} dans un fichier
de configuration \code{.Renviron}.
Consulter \href{https://api.insee.fr/catalogue/site/themes/wso2/subthemes/insee/pages/help.jag}{cette page}
pour de l'aide sur comment obtenir de tels identifiants.
}

\examples{
\dontrun{

sirets_successeurs(c("30070230500040", "30137492200120", "30082187300019"))
}
}
